FIX: Test::runTests(): Accept Callback instead of Reference
Since invoked tests might also change the logger module name of the current thread, they can have side-effect on the test framework logging.
Therefore, instead of invoking functions of the passed Console reference, we shall always invoke CConsole::getConsoleInstance() to acquire the Console reference because that also updates the current logger module name!
To keep convenience and to keep the flexibility for the caller, we now use a callback function, so the caller passes a callback function which returns the required Console reference for us but also always invokes CConsole::getConsoleInstance(), to keep the current logger module name correct even if invoked tests changed it.